refactor: ImageInfo를 global.storage.dto.ImageResponse로 이동 (#142)
