+++ src/main/java/server/MATE/domain/test/dto/response/TestDetailResponse.java
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import server.MATE.domain.test.entity.Category;
 import server.MATE.domain.test.entity.Test;
 import server.MATE.domain.test.entity.TestCategory;
+import server.MATE.global.storage.dto.ImageResponse;
 
 import java.util.List;
 
@@ -13,13 +14,13 @@ public record TestDetailResponse(
         String title,
         List<String> categories,
         @Schema(description = "이미지 키·URL 쌍 목록 (key: 수정 요청용, url: 렌더링용, 30분 만료)")
-        List<ImageInfo> images,
+        List<ImageResponse> images,
         Integer reward,
         String description,
         String serviceName,
         String serviceDescription
 ) {
-    public static TestDetailResponse from(Test test, List<ImageInfo> images) {
+    public static TestDetailResponse from(Test test, List<ImageResponse> images) {
         List<String> categoryCodes = test.getCategories().stream()
                 .filter(testCategory -> testCategory.getDeletedAt() == null)
                 .map(TestCategory::getCategory)
